pyproject.toml
[tool.poetry]
name = "eawork-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
bcrypt = "^3.2.0"
sentry-sdk = "^1.3.1"
psycopg2-binary = "^2.9.1"
python-environ = "^0.4.54"
python-dotenv = "^0.19.0"
python-dateutil = "^2.8.2"
python-gitlab = "^2.10.1"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
Pillow = "^8.3.2"
ddtrace = "^0.51.2"
dnspython = "^2.1.0"
tldextract = "^3.1.2"
stripe = "^2.61.0"
simple-salesforce = "^1.11.4"
zeep = "^4.1.0"
google-cloud-bigtable = "^2.3.3"
google-auth = "^2.1.0"
google-cloud-core = "^2.0.0"
google-cloud-bigquery = "^2.28.1"
sqlalchemy-bigquery = "^1.2.0"
google-cloud-bigquery-storage = "^2.9.1"
pyarrow = "^5.0.0"
Django = "^3.2.6"
django-allauth = "^0.45.0"
django-adminutils = "^0.0.10"
django-modeladmin-reorder = { url = "https://github.com/bloodwithmilk25/django-modeladmin-reorder/archive/6cc5060dec2fe1e894c3ba17aeb703b56737c3e8.zip" } # for django 3.2 support
django-ninja = "^0.18.0"
django-import-export = "^2.5.0"
django-enumfields = "^2.1.1"
django-solo = "^1.1.5"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.7.2"
drf-yasg = "^1.20.0"
django-modeltranslation = "^0.17"
django-auditlog = "=1.0a1" # version update had broken migrations
django-json-widget = "^1.1.1"
django-extensions = "^3.1.3" # mostly for manage.py shell_plus
ipython = "^7.26.0"
django-add-default-value = "^0.9.1"
django-colorfield = "^0.4.2"
django-admin-sortable2 = "^1.0"
django-nested-inline = "^0.4.4"
django-tinymce = "^3.3.0"
django-storages = "1.11.1" # 1.12+ pobably needs an update of boto3 & its reqs, otherwise throws an ssl error on upload
boto3 = "^1.18.38"
django-object-actions = "^3.0.2"
django-fieldsignals = "^0.7.0"
django-select2-admin-filters = "^0.1.4"
django-select2 = "^7.7.1"
django-cors-headers = "^3.10.0"
django-model-utils = "^4.2.0"
django-safedelete = "~1.1.0" # Version 1.2 has multiple issues: https://github.com/makinacorpus/django-safedelete/issues/206
dj-rest-auth = "^2.1.11"
celery = "^5.2.3"
redis = "^4.1.0"
gevent = "^21.12.0"
cffi = "^1.15.0"
plaid-python = "^8.10.0"
Faker = "^8.11.0"
django-anymail = { extras = ["mailgun"], version = "^8.5" }
html2text = "^2020.1.16"
celery_once = "=3.0.1"
django-cache-memoize = "^0.1.10"
requests = "^2.27.1"
MLB-StatsAPI = "^1.4.2"
django-jsonform = "^2.10.0"
tweepy = "^4.10.0"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
black = { extras = ["d"], version = "^22.1.0" }
model-bakery = "^1.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py39']
line-length = 97

[tool.isort]
profile = "black"
force_single_line = true
line_length = 97
lines_after_imports = 2
